feat: implement change password service, reduce tight coupling in application and infrastructure layer, improved error handling

+++ src/auth/application/auth_use_cases.py
@@ -1,8 +1,14 @@
+from datetime import datetime
 from src.auth.domain.services.auth_services import AuthService, UserService
 from src.auth.domain.repositories.user_repo import IUserRepository
 from src.auth.domain.entities.users_entity import UserRegister, UserStore
-from fastapi import HTTPException
-from datetime import datetime
+from src.auth.domain.entities.users_entity import UserOut
+from src.shared.domain.exceptions import (
+    UserNotFoundException,
+    AuthenticationFailedException,
+    InvalidUpdateException,
+    UserAlreadyExistsException
+)
 
 
 class AuthUseCases:
@@ -12,40 +18,48 @@ class AuthUseCases:
         self.user_services = user_services
 
     def signup(self, user: UserRegister) -> dict:
-        try:
-            now = datetime.utcnow()
-            self.user_services.validate_user(user)
-            hashed_password = self.auth_services.hash_password(user.password)
-            user.password = hashed_password
-            user = UserStore(**user.dict(), created_at=now)
-            return self.user_repo.create(user.model_dump())
-        except ValueError as e:
-            raise HTTPException(
-                status_code=400, detail=f"Invalid user data: {str(e)}.")
-        except Exception as e:
-            raise HTTPException(status_code=500, detail=str(e))
+        if self.user_repo.read(user.username):
+            raise UserAlreadyExistsException(
+                f"User with username {user.username} already exists")
+        now = datetime.utcnow()
+        self.user_services.validate_user(user)
+        hashed_password = self.auth_services.hash_password(user.password)
+        user.password = hashed_password
+        user = UserStore(**user.dict(), created_at=now)
+        return self.user_repo.create(user.model_dump())
 
     def login(self, username: str, password: str) -> dict:
         user = self.user_repo.read(username)
-        print(user)
-        if not user:
-            raise HTTPException(status_code=401, detail="Invalid credentials.")
-        if not self.auth_services.verify_password(password, user["password"]):
-            raise HTTPException(status_code=401, detail="Invalid credentials.")
+        if not user or not self.auth_services.verify_password(password, user["password"]):
+            raise AuthenticationFailedException(
+                "Incorrect username or password")
         token = self.auth_services.create_access_token({"username": username})
-        return {"token": token, "token_type": "bearer"}
+        return {"access_token": token, "token_type": "bearer"}
 
-    def delete_user(self, username: str) -> dict:
+    def delete_user(self, username: str, current_user: UserOut) -> dict:
+        if not current_user.admin:
+            raise AuthenticationFailedException("Unauthorized")
+        if not self.user_repo.read(username):
+            raise UserNotFoundException(
+                f"User with username {username} not found")
         return self.user_repo.delete(username)
 
-    def change_password(self, username: str, password: str, new_password: str) -> dict:
+    def change_password(self, current_user: UserOut, password: str, new_password: str) -> dict:
+        username = current_user.username
         user = self.user_repo.read(username)
-        if not user:
-            raise HTTPException(status_code=401, detail="Invalid credentials.")
         if not self.auth_services.verify_password(password, user["password"]):
-            raise HTTPException(status_code=401, detail="Invalid credentials.")
+            raise AuthenticationFailedException("Incorrect password")
         hashed_password = self.auth_services.hash_password(new_password)
-        return self.user_repo.update(username, {"password": hashed_password})
+        data = self.user_repo.update(username, {"password": hashed_password})
+        return {
+            "status": "ok",
+            "message": "Password changed successfully",
+            "username": data["username"]
+        }
 
-    def update_user(self, username: str, update_data: dict) -> dict:
+    def update_user(self, current_user: UserOut, update_data: dict) -> dict:
+        username = current_user.username
+        if "password" in update_data.keys():
+            raise InvalidUpdateException(
+                "Cannot update password. Use reset password instead.")
         return self.user_repo.update(username, update_data)
